+++ pomdol/src/app/_lib/registry.tsx
