+++ pomdol/src/app/_styles/globalStyle.ts
