+++ pomdol/src/app/_styles/style.d.ts
