+++ pomdol/src/app/layout.tsx
@@ -1,4 +1,4 @@
-import StyledComponentsRegistry from '@/lib/registry';
+import StyledComponentsRegistry from './_lib/registry';
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
